resolving PR feedback
Including integration test helpers and updating errors handled when installing apt packages
